Udated face recognition to check for frame

+++ gen2-face-detection/utils/utils.py
@@ -22,7 +22,7 @@ def draw(img: np.ndarray, bboxes: np.ndarray, landmarks: np.ndarray, scores: np.
         color = (0, 255, 0)
         thickness = 2
         for idx in range(bboxes.shape[0]):
-            bbox = bboxes[idx].astype(np.int)
+            bbox = bboxes[idx].astype(np.int16)
             cv2.rectangle(img, (bbox[0], bbox[1]), (bbox[0]+bbox[2], bbox[1]+bbox[3]), color, thickness)
             cv2.putText(img, '{:.4f}'.format(scores[idx]), (bbox[0], bbox[1]+12), cv2.FONT_HERSHEY_DUPLEX, 0.5, (255, 255, 255))
 
@@ -38,7 +38,7 @@ def draw(img: np.ndarray, bboxes: np.ndarray, landmarks: np.ndarray, scores: np.
             (  0, 255, 255)  # mouth left
         ]
         for idx in range(landmarks.shape[0]):
-            face_landmarks = landmarks[idx].astype(np.int)
+            face_landmarks = landmarks[idx].astype(np.int16)
             for idx, landmark in enumerate(face_landmarks):
                 cv2.circle(img, (int(landmark[0]), int(landmark[1])), radius, color[idx], thickness)
     return img
